Add two new wave generators.
'1b Noise' is a periodic one-bit noise generator using the same algorithm as the SN76489 chip present in a BBC Micro.

'Buzz' is a periodic noise generator using a similar algorithm that has a shorter period. It originated as a mistake, but it makes a nice digital buzz/clang sound, so I kept it.